app/cmd/client/ui/requirements.txt
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
blinker==1.8.2
capstone==5.0.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colored-traceback==0.4.2
cryptography==43.0.1
customtkinter==5.2.2
darkdetect==0.8.0
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-cli==0.0.5
Flask==3.0.3
future==0.18.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
intervaltree==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
numpy==1.22.3
opencv-python==4.6.0.66
packaging==24.1
paramiko==3.5.0
pefile==2022.5.30
Pillow==9.1.1
plumbum==1.8.3
psutil==6.0.0
psycopg2==2.9.10
pwntools==4.13.0
PyAutoGUI==0.9.53
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pyelftools==0.31
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyJWT==2.10.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytweening==1.0.4
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ROPGadget==7.4
rpyc==6.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.13.1
typing_extensions==4.12.2
unicorn==2.1.0
unix-ar==0.2.1
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
Werkzeug==3.0.4
zstandard==0.23.0
